Rework prerender browsertests to use base::RunLoop.
Remove most of the places where we content::RunMessageLoop() and then elsewhere
base::MessageLoop::current()->Quit(). Too many parts of the tests involve
spinning the message loop and waiting for one of three or four events to
happen, some complex conditions on some of the events when we want AND instead
of OR.

The primary change is a TestPrerender object vended out by the
TestPrerenderFactory. This object is a PrerenderObserver and holds a non-owning
reference to the TestPrerenderContents. Its lifetime is not tied to the
TestPrerenderContents and is managed by the caller. TestPrerender exposes
methods to wait for various events which replaces the content::RunMessageLoop
calls.

The hope is that this will make it easier to write tests which create multiple
prerenders and allow us to unbreak and reenable tests disabled because the old
wait conditions clobbered each other.

BUG=307592, 249179
TEST=Prerender*.*

Review URL: https://codereview.chromium.org/137263002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@245100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c