Datadump on sync_internals output REDACTED instead
When the user request sync events dump on the sync_internals page,
he can check a box to say if he wants to dump also the node content.
If he doesn't check it, then the node is not stringyfied and it returns
a "[Object object]" message.

This changes this behaviour and replaces the message with a "REDACTED",
that looks better on the .csv file and makes it clear that the user
requested this option, insted of being a stringyfier or any other bug.

BUG=333365

Review URL: https://codereview.chromium.org/138563003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@245173 4ff67af0-8c30-449e-8e8b-ad334ec8d88c