Support displaying the avatar label on the right side.
The avatar label should always be displayed on the same side
as the avatar button. Currently, the avatar label is always
displayed on the left, but the avatar icon can also be
displayed on the right. This CL moves the avatar label adds
support for displaying the avatar label on the right and moves
the avatar label to the right if the avatar icon is also on
the right.
A screenshot of the placement on the right can be found in the bug description.

BUG=332392
TEST=unit_tests

Review URL: https://codereview.chromium.org/127253004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@245503 4ff67af0-8c30-449e-8e8b-ad334ec8d88c