In-app messaging to communicate the imminent 32-bit deprecation to 32-bit-only users.
In Google Chrome-branded builds, show an infobar at startup to users with
32-bit-only systems, and add a message to the about page for those users
advising that automatic updates will soon stop because the system
configuration is about to become unsupported.

A provision is made to change this message to an "end-of-the-line" message for
the final 32-bit-compatible build. When this switch is flipped, the autoupdate
status and promotion button on the about page will be hidden. The
"end-of-the-line" message will appear there, alongside the icon normally used
for failed updates. The startup infobar text will change at that point as
well.

No observable changes for non-branded Chromium builds (since they're not
receiving updates through this mechanism) or for users with 64-bit-capable
systems (because they're not being obsoleted).

R=arv@chromium.org, thakis@chromium.org, thakis
BUG=303280,331810,331830

Review URL: https://codereview.chromium.org/125443002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@245598 4ff67af0-8c30-449e-8e8b-ad334ec8d88c