Don't DCHECK being on the UI thread in failed callback handler.
If a thumbnail request fails because the browser is shutting down. In
that case the failure may be executed after the browser UI thread
is being closed, and the CurrentlyOn(content::BrowserThread::UI)
check will fail.

So move this check to only be executed when the ProcessCapturedBitmap
result is successful, and the function is going to do any work.

R=sky
BUG=270918

Review URL: https://codereview.chromium.org/141273011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@245767 4ff67af0-8c30-449e-8e8b-ad334ec8d88c