Files.app: Hides quiet progress items after they are completed.
Previously quiet progress items which don't have progress bar were wrongly
regarded as they have a progress bar and the bar is animated. The items marked
as animated wait for the CSS transition-end event but it never happens.

This CL fix the condition to determine if an item should be animated or not.

BUG=336309
TEST=manually

Review URL: https://codereview.chromium.org/135573003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246024 4ff67af0-8c30-449e-8e8b-ad334ec8d88c