[Hotword] Adding pref to track number of times the opt-in popup is shown. Adding functions to the HotwordService and HotwordServiceFactory to find out whether the popup should be shown based on the pref. A test is included for these added functions.
This also fixes some pref accessing bugs in HotwordServiceFactory that were found when writing the test.

Note: The hook up to the extension javascript will occur in a separate CL.

Note: There seem to be two bugs for this so adding this to both for tracking.

BUG=325439,289023
TEST=HotwordServiceTest.ShouldShowOptInPopup

Review URL: https://codereview.chromium.org/136113002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c