Files.app: Fix a moving progressing message from Downloads to Drive.
To move a file between different volumes, FileOperationManager uses CopyTask and
removes the original files after copying. This caused to a bug that the message
of the copy task is assigned to the move operations.

This CL adds a deleteAfterCopy flag to the constructor of CopyTask and
determines the message by referring the flag.

Along the way, this CL removes lines that is not used in FileOperationManager.

BUG=336335
TEST=manually

Review URL: https://codereview.chromium.org/143343004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246293 4ff67af0-8c30-449e-8e8b-ad334ec8d88c