Linux GTK: Always puts the Omnibox text center.
Fixes an issue that the text in Omnibox is jumping up and down while typing.

GtkTextView, used for Omnibox, is a multi-line text edit control and it's a scrollable view.  So, when the line height is greater than the Omnibox height, it's unpredictable which part of region of GtkTextView is shown, i.e. how much pixels it's vertically scrolled.  It can be top or bottom (or theoretically any other values).  This is the cause of jumping text.

The fix in this CL determines the vertical scroll position of GtkTextView respecting the cap height of the font.  So it's now deterministic.

BUG=277061
TEST=Done manually changing the theme to Classic, GTK+, and custom theme.

Review URL: https://codereview.chromium.org/143953009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246560 4ff67af0-8c30-449e-8e8b-ad334ec8d88c