mac: Fix a bug where the history swipe arrow would fail to disappear.
Previously, the arrow was being added to the window. I changed the logic to add
the arrow to the render_widget_host_view. I was unable to trace down the source
of the bug, but I suspect that there are race conditions that induce a memory
leak of the chrome_render_widget_host_view_mac_delegate, as that's the only way
for the arrow to fail to disappear.

I have been unable to reproduce the bug locally, but mehmet@ was able to
reproduce it on his machine, and this change fixes the problem.

BUG=330312
NOTRY=true

Review URL: https://codereview.chromium.org/139743013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246604 4ff67af0-8c30-449e-8e8b-ad334ec8d88c