Temporary fix for bad hit testing of tab close button
The problem is the close button ends up saying it has a hit test mask,
but returns an empty hit test mask. The code was checking for
containment, when it should be checking for intersection.

My fix isn't ideal, we really need to return the intersection, but
this should be good for now.

BUG=332334
TEST=see bug
R=msw@chromium.org, sadrul@chromium.org, tdanderson@chromium.org

Review URL: https://codereview.chromium.org/134203004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@246686 4ff67af0-8c30-449e-8e8b-ad334ec8d88c