Introduce SigninManagerBase::Observer.
SigninManager will move into the core code of the signin component, and thus
cannot send notifications (which are a //content-based concept). This CL
introduces SigninManagerBase::Observer and changes AccountReconcilor to
observer SigninManagerBase rather than listen for notifications. Subsequent CLs
will change the other clients of the SigninManager notifications. Note that the
Observer is defined at the SigninManagerBase level even though it is
SigninManager that sends calls to the observers; this is done in order to
enable cross-platform code that only knows about SigninManagerBase to add and
remove observers of signin.

This CL also changes the SigninManager unittest to test the observer-based flow
rather than the notifications-based flow. 

TBR=thakis

Review URL: https://codereview.chromium.org/136723002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247042 4ff67af0-8c30-449e-8e8b-ad334ec8d88c