Fix dangling pointer in Isolated FS MessageFilter
In CrxFs API, PepperIsolatedFileSystemMessageFilter used to
const-ref to BrowserPpapiHost via GetProfileDataDirectory.  But when
plugin is dead or dying, it's possible that BrowserPpapiHost has been
destroyed but MessageFilter is still alive, since the former is not
ref-counted.  When that happens, PepperIsolatedFileSystemMessageFilter
holds a dangling reference.

MessageFilter now stores profile directory itself.

TEST=out/Debug/browser_tests --gtest_filter="PPAPITest.*"
BUG=N/A

Review URL: https://codereview.chromium.org/147303002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247387 4ff67af0-8c30-449e-8e8b-ad334ec8d88c