Adjust pretty-printing for "Google Drive" path in download option for CrOS.
This is the Google Drive counterpart of what I have done
for local "Downloads" folder in crrev.com/232055.

For multi-profile support, we are rewriting paths that were
unique inside a profile but collides between multiple profiles.
In the Drive folder case, it used to be "/special/drive" but
it will become "/special/drive-<some-id-for-profile>".

BUG=336123

Review URL: https://codereview.chromium.org/148093002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c