Don't create the UserPolicySigninService at profile creation time on iOS.
This service is used to trigger a policy registration at Profile creation time
on other platforms but this behavior isn't used on iOS.

Creating it too early on iOS also leads to a crash, because on this platform
the refresh token is available at this stage and the SigninManager will attempt
to refresh tokens before the ProfileIOData is ready. This isn't an issue on
the other platforms because the OAuth2TokenService hasn't been loaded yet.

BUG=275292
R=dconnelly@chromium.org

Review URL: https://codereview.chromium.org/143323015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247441 4ff67af0-8c30-449e-8e8b-ad334ec8d88c