Clarify and separate the concept of "Drive path" and "Filesystem relative path".
- RelativeFileSystemPath (also called VirtualPath) is the path appended
in the filesystem URL, after .../external/. So it looks like
"drive-xxx/root/foo/bar" or "Downloads-xxx/foo/bar".
Note that in the Drive case, it begins with the mount point name,
that differs from profile to profile.

- DrivePath is the path in the hierarchy we manage in drive::ResourceMetadata.
It is like "drive/root/foo/bar" and always begins with "drive".

Every path comes from drive::FileSystem is in DrivePath.
Every path that will be the part of FileSystem API must be in RelativeFileSystemPath.
We need proper conversion for the bridging codes.

BUG=337095

Review URL: https://codereview.chromium.org/147993003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247561 4ff67af0-8c30-449e-8e8b-ad334ec8d88c