Fix UMA recoding code for FileBrowser.ViewingFileType.
In r239446 I carelessly inserted '.ico' to the array but
I should have not. The array is for indexing enum of metric
stat that need to match with remote side config.

Let me just revert the addition. I think we are not so eager
to count '.ico''s popularity.

BUG=339023
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/130803008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@247826 4ff67af0-8c30-449e-8e8b-ad334ec8d88c