aura: Remove event-dispatch methods from WindowTreeHostDelegate interface.
Update the aura::WindowTreeHost implementations to also be ui::EventSource, and
remove the event-dispatch related methods from the WindowTreeHostDelegate
interface. The implementation of the WindowTreeHostDelegate interface related to
event-dispatch in RootWindow uses ui::EventProcessor for dispatching events
anyway, and so it is a good time to remove this redundant code.

BUG=318879
R=ben@chromium.org

Review URL: https://codereview.chromium.org/147203004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@248154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c