[Mac] Fix crash when the Gaia embedded view is open and Chrome is force-quit
The crash happens because when Chrome is quit with Command-Q and the bubble
is open, although a "browser is closing" notification is sent, Cocoa doesn't
get to clean up the UI (the bubble's -dealloc isn't called). If the
embedded view is open, this means that we're leaving a dangling WebContents
renderer around, and the ProfileDestroyer is upset.

To fix this, always reset the WebContents when the browser is closing. This
won't make a difference in the normal case.

BUG=339759
TEST=Open Chrome with the --new-profile-management flag. Click on the
avatar button of a non-signed in profile. Choose "Log into Chrome".
Press Command-Q. You shouldn't crash.

Review URL: https://codereview.chromium.org/132483011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@248381 4ff67af0-8c30-449e-8e8b-ad334ec8d88c