Accept smaller icons for the extension uninstall dialog for bookmark apps.
The extension uninstall dialog currently only accepts 128x128 or larger
icons. Bookmark apps may not have this icon size. Since the 48x48 icon is
shown in the launcher, bookmark apps should use that icon in the extension
uninstall dialog to prevent the case of having an icon in the launcher, but
not in the uninstall dialog.

BUG=318607

Review URL: https://codereview.chromium.org/153393002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@248863 4ff67af0-8c30-449e-8e8b-ad334ec8d88c