Set/override WebContents preferred size during tab capture.
During tab capture, the preferred size of a WebContents will be the video resolution of the capture.  In upcoming changes, views will use this as a target for sizing a fullscreen widget embedded within a tab; meant to prevent additional detrimental scalings of the content.

The normal "contents preferred size" is provided by blink::WebView and propagated from the renderer to browser process, ultimately into WebContentsImpl.  Currently, the only consumer of the value is the Cocoa browser window controller, where it is an input to the calculation that determines the "standard size" of the window (when the user clicks the window's zoom button).  Thus, for this single existing use case, it is both reasonable and desirable for tab capture to override the WebContents' preferred size.

BUG=256870
R=imcheng@chromium.org, joi@chromium.org, sky@chromium.org
TBR=shishir@chromium.org

Review URL: https://codereview.chromium.org/148583011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c