Files.app: Fetch drive metadata when entries are retrieved.
Previously DirectoryContentsDriveOffline is implemented by traversing the drive
directory tree and fetching the metadata of files. This caused to cache metadata
of drive files. But the implementation is replaced with
fileBrowserPrivate.searchDriveMetadata by crrev.com/201891. So we need to fetch
drive metadata explicitly now.

BUG=341395
TEST=manually

Review URL: https://codereview.chromium.org/150133005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c