Fix leak in PrerenderManager::MarkWebContentsAs*
If a mark happens in the last navigation a WebContents ever does, we leak an
entry in the hash table on PrerenderManager. Instead, move all the state
machines and logic to PrerenderTabHelper. Remove prerender_manager_unittests.cc
as they are no longer relevant. Instead, add tests to PrerenderBrowserTest
which assert that the correct PPLT events get emitted in the relevant cases.

BUG=334781
TEST=PrerenderBrowserTest.PrerenderPage
     PrerenderBrowserTest.PrerenderPageCrossDomain
     PrerenderBrowserTest.MatchCompleteDummy
     PrerenderBrowserTest.PrerenderDeferredImage
     PrerenderBrowserTest.PrerenderPPLTNormalNavigation

Review URL: https://codereview.chromium.org/131613006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249441 4ff67af0-8c30-449e-8e8b-ad334ec8d88c