[OriginChip] Hide/show the origin chip as the Omnibox gains/loses focus.
The ultimate implementation may be rather more complicated, but this simple approach covers nearly all of the currently specified behaviors and should be a good starting point for everyone to play around with.

Also add a check for editing in progress when the browser view sets focus to the location bar so that Ctrl-l/Alt-d won't clobber any edits in progress.

BUG=338563

Review URL: https://codereview.chromium.org/143043004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249555 4ff67af0-8c30-449e-8e8b-ad334ec8d88c