kiosk: Enable Ctrl +/-/0 zoom content accelerators for app windows.
We do not enable these for app windows in general, because the user
can already adjust the zoom using the page view settings. However, 
this is not possible in ChromeOS Kiosk mode. This patch provides a way
to zoom app windows in ChromeOS Kiosk mode.

It's a minimum subset of mlamouri's
  https://codereview.chromium.org/97003002/ to fix the bug.

BUG=341235

Review URL: https://codereview.chromium.org/155743002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249666 4ff67af0-8c30-449e-8e8b-ad334ec8d88c