Relanding WebRTC browser tests will now render during tests
WebRTC browser tests will now render during tests.

This is a speculative fix to try to fix the video quality test timeouts
on the FYI bots. I noticed the browser tests did not run with pixel
output enabled, which means they don't really render. Not sure how it
is related to the gUM failures in the bug, but this will at least make
it easier to see what's going on when it fails

TBR=xians@chromium.org
BUG=340609

Review URL: https://codereview.chromium.org/132373009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249688 4ff67af0-8c30-449e-8e8b-ad334ec8d88c