Bluetooth: avoid sending cancel after accept or reject
After clicking the Accept or Reject buttons, the pairing dialog is closed
which results in a CancelPairing method call being sent after the reply
to the original method call.

This isn't harmful right now, but could be a bug if BlueZ changes, so fix
it now.

BUG=none
TEST=verify with d-bus monitor that only a method reply is sent

Review URL: https://codereview.chromium.org/148643003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c