Split PrefHashStore into PrefHashStore and PrefHashStoreTransaction.
All a PrefHashStore can now do is hand off PrefHashStoreTransactions.

PrefHashStoreTransactions are used to determine the scope of a series of updates to the PrefHashStore and do one-time computation at the end when required.

This change is aimed in particular at updating the hash_of_hashes only once at the end of a series of updates to multiple hashes rather than recomputing it every time a hash is stored in that serie.

BUG=340540
R=bauerb@chromium.org, erikwright@chromium.org

Review URL: https://codereview.chromium.org/153433003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@249843 4ff67af0-8c30-449e-8e8b-ad334ec8d88c