Revert 249996 "Add key based scrolling on magnified screen for k..."
Suspecting breaking "Win7 Tests (dbg)(4)" and "Win7 Tests (dbg)(5)" due to failed OutOfProcessPPAPITest.Fullscreen tests. The level of confidence in this revert is low but the blame list is short. This is
the CL that touches UI and the bot was consistently green before.
Sorry for reverting, please try to land later when the waterfall is green.
 
> Add key based scrolling on magnified screen for kiosk mode.
> 
> * Add MagnifierKeyScroller event handler that interprets
>   click and hold action.
> * kAshEnableMagnifierKeyScroller to test on desktop.
> * Modified EventGenerator to generate native event for key press/release.
> 
> BUG=340729
> TEST=MagnifierKeyScrollerTest
> 
> Review URL: https://codereview.chromium.org/155493002

TBR=oshima@chromium.org

Review URL: https://codereview.chromium.org/135213004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@250009 4ff67af0-8c30-449e-8e8b-ad334ec8d88c